fix: extract connection list tabs component
