+++ config-ui/.gitignore
@@ -19,7 +19,7 @@ cypress.env.json
 # misc
 .DS_Store
 *.pem
-
+	
 # debug
 npm-debug.log*
 yarn-debug.log*
@@ -35,5 +35,5 @@ yarn-error.log*
 .vercel
 
 # ide & workspace files
-.nova
+/.nova
 *.sublime*
